Move to gitlab autodeploy image 1.1.0
The main change is upgrading kubernetes from 1.13.12 to 1.15.12, full
changes are here:

https://gitlab.com/gitlab-org/cluster-integration/auto-deploy-image/compare/v1.0.7...v1.1.0

There was no particular reason for doing this, apart from that it
seems good to take on new versions where we can.

I believe we can't move to version 2.x of this image as that would
cause an upgrade to helm 3 which would need us to manually do some
migration:

https://docs.gitlab.com/ee/topics/autodevops/upgrading_auto_deploy_dependencies.html#upgrade-deployments-to-the-v2-auto-deploy-image

(The instructions on that page almost certainly won't work for us as
our .gitlab-ci.yml overrides a lot of the default one.)

We probably should attempt this sometime as helm2 is no longer
supported.