Use MTLS aliases from discovery whenever using MTLS constrained tokens
The MTLS version of the token endpoint should always be used in FAPI
as MTLS sender constraining of access tokens is always used. Apply
the MTLS aliases for private_key_jwt as well.

(The code could potentially be simplified a little now MTLS aliases are
used in all cases in FAPI, though in OIDCC it remains conditional so
some of the complexity would remain.)

closes  #883