Copy FAPIRWID2 tests to FAPI1AdvancedFinal
Named is as per https://bitbucket.org/openid/fapi/issues/362/language-to-use-on-the-certification-pages

This was done by:

* renaming the existing test package in Intellij

* undoing the deletions of the old locations

* removing changes to the FAPICIBA tests, as they are for FAPICIBA ID1 and
the small parts of FAPI code they import should continue to be shared
with the ID2 tests.

The change was done like this so that all the subclasses of the test modules
are also duplicated.

The reason for adding complete new versions of the tests is that:

a) Eventually the fapirw id2 ones will be completely deleted

b) All the spec references through all the test modules need to be changed
from ID2 to Final

Part of #885