Adding support for filtering out test plans
We have forked this repository to use for functional testing of Open Banking Brasil. This requires that we surface only relevant test plans in the UI. This change utilises the profile of the @PublishTestPlan annotation to allow us to optionally configure the subset of profiles which will be returned by VariantService, inside application.properties. If this value is not configured, then all present test plans will be presented. This allows us to take advantage of the functionality without impacting any other instances, whilst minimising our own efforts when merging in upstream changes