Replace tabs with spaces in html
Our html is currently a mix of spaces & tabs, meaning indents often
look weird in git diffs etc. The editorconfig has been set to use
spaces only for a while but there's a lot of historical tabs.

Just remove them all with:

find src -name '*.html' | xargs tab24space

where tab24space is:

 #!/bin/bash
 for i in $*; do
   expand -t 4 $i > $i.new || exit 1
   mv -f $i.new $i || exit 1
 done

Some of the indenting may be incorrect now as I'm sure some was
assuming 4 character tabs, but we can fix that as we see it.