Brazil: Create DCR 'happy flow' test
Obtains an SSA from the directory then registers at the bank and checks
the client works.

Obtaining the SSA means we need to do discovery and client credentials
grant at the directory, this has been done using the existing test
conditions and mostly mapping the environment keys to avoid the
directory values overwriting the ones for the main part of the test.

The DCR mostly follows RFC7591, but some of the fields in the SSA do
not follow the standard - however we do very little processing SSA.

The same MTLS certificate is used to authenticate to the directory as
a client, to authenticate at the DCR endpoint and for client
authentication/certificate bound access tokens to the system under
test.

This will be a separate column on the certification page I believe so
has been done as a new test plan rather than generally adding DCR as
an option to the FAPI tests. The test plan will eventually have a few
more tests.