Refactor GetDynamicClientConfiguration into two conditions
It had a non-obvious side effect of extracting client_name into the
root of the environment, put this in it's own condition.

Further renaming in the next commit will rename the main condition and
it's output environment variable to make a bit more sense, as the naming
doesn't really reflect how it's actually used.