Rename GetDynamicClientConfiguration to StoreOriginalClientConfiguration
This seems to better reflect what it actually does.

Also do the same for the Client2 version.

Note that both functions store the function to the same environment
variable, and none of the test modules seem to map that key, so the
client 1 values aren't available once client 2 starts being processed.