Perform some extra validation on JWTs
A support query asked about this odd error when extracting an encrypted
id_token:

Invalid unsecured/JWS/JWE header: Invalid JSON: Unexpected token
??u?m?? at position 7.

The root cause turned out to be that the payload had not been signed
before it was encrypted, but the error is shockingly unhelpful.

I'm not entirely sure why we got this error, but it's clear that nimbus
(or something underlying it) is happy to base64url decode something that's
very clearly not valid base64url. So we had some extra validation that
a jwt contains valid base64url before we pass it over to nimbus. It
probably doesn't catch every instance, but it's a good start that should
cover the most common cases.