Upgrade to bouncycastle 1.64 (from 1.58)
This was part of investigating why some fields in a newer-format
Brazil TLS certificate subjectdn weren't being decoded into the same
format as openssl x509 -in foo.pem -text would show.

Unfortunately it didn't help, but it seemed to sensible to move up to
as late a version as we can anyway.

org.bouncycastle.crypto.tls is unfortunately deprecated (and completely
removed in later versions than 1.64), so suppressions have been added
for the deprecations. The replacements are in org.bouncycastle.crypto.tls,
and I've added the relevant dependency to pom.xml, but they are not yet
used anywhere as there's a number of API changes.