FAPI: Add a warning about mtls aliases + client assertions
I suspect this code should probably be considered wrong, but I'm not
sure this is a good time to make a behaviour change, and it would
probably be wise to seek wider consensus before doing so.

The value we are using is arguably acceptable, at least when the
assertion is being sent to the token endpoint and hence is seems
acceptable for the 'aud' to be the url the assertion is actually
being sent to.