FAPIBrazil: Add CI tests
We ran into problems trying to automate the raidiam mock bank
(selenium's htmlunit doesn't seem to like it's javascript) and there's
no way to skip it / run headless. So add configs against Authlete
instead.

This is just an initial set; it'd probably be good to test MTLS client
auth and a few other combinations.