Fix alias handling when multiple users involved
If two users are trying to run tests with the same alias, we do not
add the usual 'test has stopped because another test has claimed' the
alias message, because the code that adds this only does so if the old
test is owned by the current user.

But the new user does claim the alias from the old user anyway.

This is causing a lot of confusion in Brazil currently as many people
use the 'obbsb' alias from the example configuration.