BrazilDCR: Fix problem if server under test has no mtls endpoint aliases
AddMTLSEndpointAliasesToEnvironment had inconsistent behaviour -
if the server has an mtls_endpoint_aliases object, it puts all the
endpoints into the environment root, overriding them with the mtls
alias if applicable.

However if the server has no mtls_endpoint_aliases, nothing would be put
into the environment. In the particular case in Brazil this left the
directory's MTLS aliases in the environment when they should have been
replaced by the endpoints for the server under test.

part of #914 (this should fix the actual real problem we saw there, but
there remains an issue that the mtls aliases should probably be in an
object within the environment so they can be mapped, and that the
handling doesn't work given sometimes we need to call the mtls alias
for some endpoints but not others).