Refactor CallDynamicRegistrationEndpoint + additional checks
It now works more like CallTokenEndpointAndReturnFullResponse,
i.e. allowing the test modules to perform more of the checks themselves,
but uses convertResponseForEnvironment so that the response is contained
within a single object in the environment and hence can be mapped -
meaning we can have generic conditions to check the contenttype, http
status, etc.

We are now more explicit about checking some of the results, e.g. the
201 response code and the json content type.

This also means we no longer need a separate condition for calling the
endpoint when we expect an error, and hence
CallDynamicRegistrationEndpointExpectingError has been removed.