BrazilDCR: Ensure people recreate test plans to include new tests
The newly added tests are important, ensure that people are running
them by renaming the existing test modules and adding a more explicit
error message for the case where a test module is no longer present
in the system. This means if they created a test plan on an older version
of the suite they will be forced to recreate it after the new version
is deployed.

We should probably think about a better mechanism for achieving this in
the longer term. e.g. Perhaps all test plans should be regarded as
immutable 2 weeks after they are created.