BrazilDCR: Fix issue when subjectdn contains non-ascii values
When calling the dynamic registration endpoint, the
tls_client_auth_subject_dn was ending up getting corrupted, resulting
in the client not functioning.

Investigation showed that the default StringHttpMessageConverter was
converting the string into Latin1, resulting in the corruption.

Override the converter with one with an explicit utf8 charset set.

It looks like every http call the suite makes has this same problem,
but I'll test/investigate further tomorrow and see if further fixes
are needed.