Add non-ascii test for FAPIBrazilExtractClientMTLSCertificateSubject
When looking at the recent bug with non-ascii characters for the
subjectdn being incorrect in the dynamic client registration request I
initially expected that the problem was in the extracting the value
from the certificate, so added this unit test.

It turned out the problem was CallDynamicRegistrationEndpoint, however
adding the unit test still seems worthwhile.