Fix UTF8 corruption in all cases
commit 688a2312 applied a fix to CallDynamicRegistrationEndpoint to
avoid it corrupting UTF8 - move that fix up a level to avoid corrupted
utf8 for any endpoint.

This also applies the "don't add a default Accept-Charset:" header
in all cases, which seems sensible, there doesn't appear to be any
cases where we rely on spring adding that header.

The new code is a little bit more complicated than the fixes used in
the specific endpoint calls, as we need to retain the existing default
converters as other endpoint calls use them.