FAPI1Adv: Show certification profile name in frontend
Allow test plans to expose a certification profile name by implementing
a static method on the plan.

The profile name is stored in the database at creation time, as it's possible later
changes to the suite could prevent it being calculated on the fly.

As an added benefit, this allows test plans to reject invalid
variant selections at creation time.