Disable http client automatic redirect & retry handling
A question from Filip revealed that anywhere in the suite where we
currently require a 200 response we also accept a 301/302 redirect
followed by a 200 response.

I think it most cases we don't expect a redirect, and redirects don't
make it into the test log so it's currently not possible to see if one
has occurred.

Hence disable automatic handling of redirects. It's unclear if there's
some places where we should be accepting redirects. If there are it's
probably better to handle them explicitly in our code so that they can
be fully logged.

Following similar logic, also disable automatic retry.

Related to https://github.com/panva/jose/discussions/240

closes #931