CheckServerConfiguration: Check values are urls
The check was originally only in the heart variant of this condition,
the refactoring done in
commit 511a96593ff28c7edd28b24c63f950f11837ee38 (fixing #604)
retained the current behaviour of the FAPI / CIBA checks.

I can't see any reason to disable the check though, everything we're
checking should be a url, so enable the check in all cases by removing
the empty override of ensureUrl. (and also tweak an error message
to be slightly clearer.)

The unit test was for some historical reason using a non-url issuer,
but it seems that was just incorrect test data.