Force users to recreate testplans to pull in added tests
Add a rather 'quick and dirty' method to ensure that Brazil testers
will not continue using old test plans that are missing some tests.

This had a slightly wider scope than necessary, as the tests added in
this version of the suite are Brazil specific and this forces all FAPI
testers to recreate their plan - however we generally don't want users
to keep on using old plans as their first test results may no longer
be valid if they take weeks to fix their product to pass all tests.