AbstractPermissionsCheckingFunctionalTestModule: Remove weird code
The Customer-Personal-api-wrong-permissions-test was not going to
'finished' state.

I can't really understand what this code was trying to accomplish -
I'm fairly certain 'waitForBrowserToReturn()' would only wait for
selenium works to complete, and even that it does in the background.
However as it ran it in a finalisation task, it prevented the real
finalisation task from running that would set the test to completed
(and do any other necessary cleanup).

Removing that weird code and just having a normal 'fireTestFinished()'
appears to result in all tests in this plan passing, so go with that.