Add payments FAPI test
This aim here is to make sure the payments API is getting the same
testing it would if the bank had run the FAPI tests against the
payment APIs. (Various banks, e.g. phase 2 ones, have definitely
not done this, and having those checks present in the functional
payments tests is the best way to make sure they do.)

This uses the FAPI1Adv happy path test module, with tweaks as
the functional tests don't expect the user to provide 'scope' and
expect the user to provide the API base rather than the payments
endpoint.