Fix for exception in oidcc-registration-jwks-uri
If the authorization server tries to fetch the jwks during the dynamic
registration call, it aborts the test.

Fetching it at this point appears to be legal behaviour, so introduce
a workaround that allows it.

I also changed the url for the client jwks, so that the workaround
only applies in this one case. (As I believe the detection of this error
was originally for client tests to make it clear the client can't fetch
the server jwks until the test has finished it's initial setup.)

Tested locally by adding a sleep in CallDynamicRegistrationEndpoint and
using curl to call the jwks endpoint.