Explicitly convert nimbus JWK to JSON in logging
This is a workaround for #662, albeit it's probably also a fairly
decent thing to do in the long term - even if mongo didn't end up
throwing a stackoverflow when logging a JWK that has an x5c, it does
ending up logging a huge object as the JWK contains a parsedX5c entry
that contains a lot of byte arrays.

I'm not entirely clear if we should ever really allow people to just
log java objects, it may make sense to just throw an exception if
we get an unexpected java object, but I don't yet entirely follow how
the mongodb converters are intended to work.