Incorporate running FAPI OP tests against RP tests into CI
This allows us to partly check that our relying party and identity
provider tests are consistent with each others. Currently only the
FAPI1-Adv-Final tests are tested, and only the 'plain_fapi' ones,
and a single Brazil accounts test.

It would be good to try and cover the OIDCC tests and further
UK/AU/Brazil open banking ones in due course.

It introduces several new syntaxes to the run-test-plan.py test plan
name field:

':' - used to specify we only want to run on module in a test plan;
for the RP tests this is the main happy flow. For OP tests we use
'fapi1-advanced-final-ensure-request-object-with-multiple-aud-succeeds'
which is a simple 'happly flow' test that only uses a single client
and doesn't call the resource server endpoint multiple times.
(Technically it might've been neater to create the test modules
directly outside of a testplan, but this would've involved much more
code change to run-test-plan.py to allow the data structures to hold
test modules run outside of plans.)

'{}' - used to specify that an OP test plan / module / config should
be run when the client test enters 'WAITING' status - done by simply
having run_test_plan call back into itself, and ensuring the parent
call returns the results from the child as well as it's own.

run-test-plan.py also gets a new feature when processing the test
configuration files - {BASEURL} can now be used in the json to refer
to the test suite external host, which is necessary e.g. to be able to
specify the discoveryurl to point at the RP tests.

The new lines in run-tests.sh are pretty ugly; moving to a config file
(similar to as is used in the OIDCC client tests) might improve
readability.

There are a couple of failures/warnings occuring currently which have
been added to the 'expected-failures' file.