FAPIBrazilRegisterClient: Generate only alphanumberic client ids
Both ; and ' ' where causing issues when running RP tests against OP
tests, as the value in the form parameters mismatched with the one
in the request object, I suspect for different reasons.

For simplicity stick with alphanumber ones (no punctuation). For
completeness we should probably check in the OP test that the server
issues only alphanumberic client ids (I suspect that is the case for
most/all) and should suggest to one of the working groups that they
limit the allowable character set for interoperability reasons.