FAPIBrazilRP: Allow the MTLS tokenendpoint as the 'aud' in client assertions
The client tests currently only allow the non-mtls token endpoint.

Both appear permissible; the general acceptable set is consider to be:

1) the token endpoint

2) the issuer

3) the url the assertion is being sent to

We already have a ticket open in this area; I've added a comment that
we need to consider this issue too:

https://gitlab.com/openid/conformance-suite/-/issues/877#note_701426648

If we wanted a strict match between the OP and RP tests here it would
seem that the FAPI RP tests should only accept the MTLS tokenendpoint
when there is one.