compare-results: Include description as part of key used
Currently the test plans for Brazil DCR with payments and accounts have
the same testplan name and variants (the only difference is the scope
in the test config), resulting in payments sometimes being compared
against accounts and vice versa. Include the description specified
in the test config to avoid that.

Also rename a badly named variable; it contains the plan name, not the
module name.