FAPIBrazilDCR: Add test that uses invalid MTLS certificate
The certificate is based on the genuine one. but doesn't validate back
to the official chain. I'm not sure exactly how good the MTLS certificate
I'm faking up here is, that might be worthy of further investigation
and we could consider adding similar tests to FAPI/FAPICIBA/OIDCC tests
that use MTLS certificates.