FAPIBrazilDCR: Verify client id/config creds returned from config endpoint
Note that we don't allow the registration access token to be rotated.
This would seem a bad thing to do for the same reason refresh tokens
shouldn't be rotated, i.e. there is a risk that the new token is not
returned and client management is irretrievably lost.

As requested from Brazil sandbox/security squad, row 18:

https://docs.google.com/spreadsheets/d/15Wu1unFOIgZy6A-jmawc2QkYWLV9SZ0W/edit#gid=1277899644