Fix fapi1-advanced-final-brazildcr-update-client-config
It was using the incorrect access token when calling the SSA endpoint
towards the end of the test. The reason for this is that the
directory access token had been overritten with one from the server
under test. So save the directory access token into a seperate
location to avoid this.

(As the raidiam mock bank can't be automated, I'd disabled the
authorization endpoint/token exchange part of the test when testing
this test so didn't see this issue.)

closes #957