FAPIBrazilDCR: Fix case where wrong dcr endpoint called
If an OP listed an endpoint only in mtls_endpoint_aliases (which I'm
at least not sure was an intended use of mtls_endpoint_aliases) the
DCR tests ending up calling the directory's registration endpoint as
the alias extraction only looked at keys present in the root of the
discovery responses.

Now we look at both sets of keys.

closes #963