Flag failures for incorrect uses of mtls_endpoint_aliases
Issue #963 revealed some very odd behavour of the suite (which is
already fixed), but the trigger was the OP listing the token_endpoint
only in mtls_endpoint_aliases, which no token endpoint in the root
of the server metadata object.

This is incorrect behaviour - as per RFC8414:

token_endpoint
URL of the authorization server's token endpoint [RFC6749].  This
is REQUIRED unless only the implicit grant type is supported.

and this incorrect behaviour is now flagged as a test failure,
replacing the previous weird behaviour of calling the directory's
token endpoint when it was intended to call the AS's.