Fix semicolons being treated as a separator in query strings
For some reason Apache's URLEncodedUtils.parse treats both ; and & as
seperators by default. Explicitly tell it to only use &

Also add some unit tests. For most cases this is simply adding tests
that try and demonstrate what the existing behaviour is. I've not checked
to make sure the existing behaviour is actually deliberate.

closes #977