Payments: Add two DCR tests
Firstly a happy flow test that ensures we can register a hardcoded
client authorised to make payments and that it can make a payment.

Then a negative test that uses a hardcoded client not authorised to
make payments, verifies it is not granted the payments scope and that
it cannot add it with a PUT to the client configuration endpoint, and
verifies it cannot obtain a 'payments' access token using the client
credentials grant.

An inelegant mechanism is use to override the alias from the test
configuration for these two test modules, so that the redirect uri
matches the ones registered for the two hardcoded clients.