FAPIBrazilRP: Test rotation of client configuration access token
Rotation of the client management endpoint access token is explicitly
permitted in Brazil and a number of banks are actually doing it.

Require the client to demonstrate a second call to the DCM endpoint,
and issue a new access token in the first call - the second call will
fail if the client does not use the new access token.

Make a corresponding update to the OP tests, mainly so the RP and OP tests
can continue to be tested against each other.