PaymentsDCR: Test two forms of subjectdn
As per RFC4514, banks should accept a subjectdn where the numeric
form of the OIDs are used, and as per Brazil DCR spec the banks should
also accept the namew form. We hence test both.

This unfortunately meant pulling in a dependency on sun.security.x509
(which is what the bouncycastle calls we were previously using called,
but they did not provide the option to pass through additional OIDs,
so we have to call the underlying API directly to do this.)

These APIs are technically internal and unstable, so we get a warning
from the java compiler for using them directly (even though we were
already indirectly using them). There doesn't seem to be a way to
disable that warning, so we have to remove -Werror to prevent the
build failing.