Resources: Add DCR/DCM tests for unauthorized clients
Refactor payments version of the test, and add a flag into the code
so it can try a PAGTO client against the resources API.

Note that this assumes a PAGTO client can't get access to 'resources'.
I believe this to be correct after a discussion with Ralph, but the roles
tables in the Brazil DCR spec implies otherwise.