Remove use of ajp
ajp has undergone some changes in springboot versions newer than we're
using, with changes coming in around version 2.2.5 that by default
require a secret to set set for ajp, and change it to bind only to
localhost by default.

Those changes were are a result of this vulnerability:

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-1938

Further investigation reveal that our cloud hosted systems (including
all of staging/demo/production and review apps) don't use ajp, and
that it seems to be recommended to migrate away from ajp:

https://people.apache.org/~schultz/presentations/ApacheCon%20NA%202020/Migrating%20from%20AJP%20to%20HTTP.pdf

Hence remove use of ajp from all our connections, which removes a
seemingly unnecessary difference between the cloud environments and
dev ones.

The move away from ajp also means that various things that ajp use
to handle automatically no longer work - in particular it dealt
with letting spring know that it was running behind a proxy and
what the real external hostname/port were. So we need to add
http headers to communicate this to spring etc. Apache adds some of
them:

https://httpd.apache.org/docs/2.4/mod/mod_proxy.html#x-headers

but not the external protocol and port number which are required by
the module spring uses:

https://tomcat.apache.org/tomcat-8.5-doc/api/org/apache/catalina/valves/RemoteIpValve.html

Without this the login-via-oidc process (which is normally disabled,
but you can enable by removing the devmode=true parameter passed
to java and added, e.g., credentials for a gitlab oauth client) failed
as it redirected back to the wrong place.

This also explains why various operations on the cloud servers like
production often redirected via http - we have never used ajp on
the cloud systems, but they also didn't pass these headers through.

Also in application.properties replace the deprecated in spring 2.2:

server.use-forward-headers=true

with the new:

server.forward-headers-strategy=NATIVE