Prevent users interrupting each others tests
Currently if two users are trying to use the same alias (which is not
recommended, but does happen in some cases e.g. if trying one of the
example configurations) then they will keep interrupting each other's
tests, and unless one gives up neither is likely to ever be able to
run a test to completion. Plus it confuses users as to what's going on
and makes it look like the test suite is randomly aborting tests.

This change brings some extra logic into the claiming of a test, so
that a user is unable to start a test if a different user is (or was
recently) using the alias.

Some timeout logic is added so that a user that wanders off whilst
performing a test will result in the alias being released after 6
minutes.

Admin user's are exempted and able to claim any alias immediately,
albeit I'm not sure if this is strictly needed.