AbstractCondition: Avoid excessive logging
Protection against a condition generating excessive logging, which
otherwise results in large amounts of data (which the frontend
struggles to process) and also puts a big load on mongodb.

A final log message is generated to record that this has happened, and
errors/warnings will continue to be logged until the limit for them
(50) is reached.

The limits picked are relatively large; they were mainly picked based
on the Brazil functional suite (which is based on our code base) which
does log some relatively large numbers of things in some of it's
conditions. They provide us with some protection in case a condition
ends up going in an infinite loop for some reason.