WIP: List items that are in the json test config but are not required
This will show items that are unneccessary, so is helpful for cleaning
up configurations to be used as examples.

It may also help a little with #986

This doesn't seem to fully work; it works when the form is first
loaded, if you switch tabs and back again the message vanishes.