schedule-test: handle values that are valid json correctly
If you insert valid json values (except numbers) into configuration
fields, they are discarded. e.g.:

false

or:

"foo"

(latter with the quotes included)

This is an odd edgecase I spotted when trying to understand the code,
the fix results in a very slight simplification to the code (but also
more importantly, easier to understand code).

Overall I think it would be better if we switched to explicitly
indicating which form fields are meant to contain what - I believe we
only have two cases at the moment:

1. String

2. JSON Object

It we had the html explicitly indicate the type for each entry field,
then for case '2' we could verify the JSON is valid before the user
submits the form, instead of them only finding out about an error when
they run the relevant test (as all invalid JSON is treated as case '1'
currently).