schedule-test: Fix ekyc_verified_claims_request not showing in form
The config form is for some reason hardwired to only support strings
at the root level of the config objects, but the new ekyc object is
at the root.

It fails because it end up searching for:

data-json-target=undefined.ekyc_verified_claims_request

The code dates back to this commit:

05ed842a9f9a03bc0bf8ba465bbc98f10a8e85ae

which explicitly added support for 'alias', by specialing casing
strings at the top level.

I can't see any reason not to just support all types of object at the
root level, so simplify the code to allow this.

closes #985