schedule-test: Pretty print json shown in form textareas
Currently when we extract values from the test json to put them into
the configuration form we remove all the nice formatting so you
end up with a jwks all on a single line in a textarea and pretty
unreadable.

Pretty print the JSON when extracting it so it's a bit more readable.