Fix issues with ValidateVerifiedClaimsResponseAgainstSchema_UnitTest
This test had been disabled due to problems getting it to pass.

The main issue was NoSuchMethodError when using JSONTokener, which
turned out to be a can of worms - the org.json library it depends on
has a 'non free' license and hence spring replaced their dependency
on org.json with an actual free library using the same namespace
but with a different API.

We probably need to remove this library; there is an alternative
suggested in pom.xml.

Once past that (by excluding the spring dependency library, it seems
like we're not using the code paths that call it or something), the
test didn't match the code it was testing, so update the test a little.